refactor(storage): extract _BaseMixin (Repository __init__ + conn)
Foundation for the per-domain mixins added in Tasks 4-9. _BaseMixin owns
__init__ (sets self._conn, self._schema) and the conn property. Domain
mixins will declare type hints for self._conn / self._schema; only
_BaseMixin assigns them.

MRO: Repository(_BaseMixin) for now; per-domain mixins added left of
_BaseMixin in subsequent tasks.
